+++ api/package.json
@@ -30,6 +30,7 @@
     "@adonisjs/repl": "^3.0.0",
     "@adonisjs/view": "^6.1.1",
     "axios": "^0.25.0",
+    "date-fns": "^2.28.0",
     "luxon": "^2.0.1",
     "mjml": "^4.10.3",
     "phc-bcrypt": "^1.0.7",
